Add atom news feed for news articles
The way the content of the `news` page is generateds has been changed to
supoport syndicating news article via an atom feed.  Rather than including flat
content from `news.md`, the page is generated from the contents of files in the
`_posts` directory. These files must be named in the form:

`YYYY-MM-DD-{title}.md`

and begin with a title section like:

---
layout: post
title: {Your Headline}
excerpt_separator: <!--more-->
---

The news articles should be formatted in markdown. When new article posts are
published they will imediately be available for newsreaders subsribed to the
atom feed. (no pun intended, but kind of beautiful the way that works out)

Long articles can be displayed on the news page as an introductory summary by
placing a line containing `<!--more-->` at the piont they want the intoduction
to be trimmed, the full article will be available by following the article link.

Note: all previous news articles are included in a single post.

Signed-off-by: Winford <winford@object.stream>