Create symlink from v0.6.0 to stable
The documentation should always point to the latest stable release, creating
this link will allow changing the master branch soucres to point to stable.

When a new release is made this symlink should be changed to point to the new
release. Ideally a way should be found to automate these changes to the `stable`
symlink.

Signed-off-by: Winford <winford@object.stream>