Add symlink to "stable"
This adds a symlink from "stable" to the current v0.6.1 documentation.
This link can be manually updated when new releases are made so that
"stable" can always point to the latest stable release. The "stable"
link can be added as a permanent menu option to the AtomVM/docs
generated documentation as well, after this link is in place.

Signed-off-by: Winford <winford@object.stream>