Remove old pre-release directories
This is a fix to reduce the size of the "Production" branch of this repo where the GitHub Pages
that host the https://www.atomvm.net website. The production branch has far exceeded its size limit
since the 0.6.5 release, and no further updates to the AtomVM documentation have been visible
on-line since.

Removes old v0.6.0 pre-release directories and temporarily adds a symlink to the v0.6.0 release
documentation until atomvm/AtomVM#1499 is merged. Once these entries are removed from the menu
these symlinks may be removed.

Signed-off-by: Winford <winford@object.stream>